fix(rust): change header nonce from u16 to u64

+++ implementations/rust/ockam/ockam_ffi/include/ockam/vault.h
@@ -206,7 +206,7 @@ ockam_vault_extern_error_t ockam_vault_hkdf_sha256(ockam_vault_t                
  */
 ockam_vault_extern_error_t ockam_vault_aead_aes_gcm_encrypt(ockam_vault_t        vault,
                                                             ockam_vault_secret_t key,
-                                                            uint16_t             nonce,
+                                                            uint64_t             nonce,
                                                             const uint8_t*       additional_data,
                                                             uint32_t             additional_data_length,
                                                             const uint8_t*       plaintext,
@@ -231,7 +231,7 @@ ockam_vault_extern_error_t ockam_vault_aead_aes_gcm_encrypt(ockam_vault_t       
  */
 ockam_vault_extern_error_t ockam_vault_aead_aes_gcm_decrypt(ockam_vault_t       vault,
                                                             ockam_vault_secret_t key,
-                                                            uint16_t             nonce,
+                                                            uint64_t             nonce,
                                                             const uint8_t*       additional_data,
                                                             uint32_t             additional_data_length,
                                                             const uint8_t*       ciphertext_and_tag,
